Fix improper specification of vertex buffer attribute offsets.
Previous implementation passed a scalar pointer to a VkDeviceSize
into vkCmdBindVertexBuffers(). This oversight caused the program
to break when there was more than one attribute binding. There is
now a const buffer of zero-value VkDeviceSizes, which can be
passed instead, avoiding the need to create a new buffer of zeros.